create cell is working :-)

+++ prison-jdbc/src/main/java/cz/muni/fi/pv168/gui/CellsTableModel.java
@@ -63,7 +63,7 @@ public class CellsTableModel extends AbstractTableModel {
     public Class<?> getColumnClass(int columnIndex){
         switch(columnIndex){
             case 0:
-                return Integer.class;
+                return Long.class;
             case 1:
                 return Integer.class;
             case 2:
