+++ prison-jdbc/src/main/java/cz/muni/fi/pv168/gui/PrisonersTableModel.java
@@ -67,7 +67,7 @@ public class PrisonersTableModel extends AbstractTableModel {
     public Class<?> getColumnClass(int columnIndex){
         switch(columnIndex){
             case 0:
-                return Integer.class;
+                return Long.class;
             case 1:
                 return String.class;
             case 2:
